Refactor worker broadcast visibility, view compiler, and alarm tick
Consolidate the trade-offer broadcast privacy contract into a single owner
(gameBroadcastVisibility.ts), make the strip/carry/re-inject contract
type-enforced, split per-domain visibility filters out of gameStateView.ts,
and extract the GameRoom alarm decision logic into a policy module.

- A: gameBroadcastVisibility.ts now owns strip/carry/re-inject/scope
  (splitBroadcastPayload, broadcastEventStateFields, buildGameScheduleEvent,
  prepareScopableGameEvent, scopeGameEventForViewer) with one doc block; all
  emit paths consume it; triplicate "keep in sync" comments removed.
- B: ScopableBroadcastPayload + StrippedBroadcastState brand make forgetting
  to strip tradeOffers a type error; emit builders accept only prepared payloads.
- C: gameVisibilityFilters.ts holds per-domain predicates + auction/log
  redaction; gameStateView.ts is a thin toClientGameState orchestrator with the
  one sanctioned InternalGameState bridge cast.
- D: gameRoomAlarm.ts (handleGameRoomAlarmTick) holds the per-tick decision;
  alarm() is a thin shell (load context, delegate, conflict-catch, resync).

Behavior and redaction identical; all suites green.
