+++ packages/worker/src/gameStateView.ts
@@ -1,68 +1,41 @@
 import type { InternalGameState } from "@oligopoly/shared";
-import { normalizeGameState } from "@oligopoly/shared";
-import type {
-  GameLogEntry,
-  GameNegotiationThread,
-  GameState,
-  InGameHandshakeAgreement,
-  PendingAuction,
-  PendingInsiderPeek,
-  TradeOffer,
-} from "@oligopoly/validation";
+import type { GameState } from "@oligopoly/validation";
+import {
+  type ClientPendingAuction,
+  filterHandshakesForViewer,
+  filterNegotiationThreadsForViewer,
+  filterTradeOffersForViewer,
+  type PersistedGameState,
+  redactPendingAuctionByMode,
+  type ViewerMode,
+} from "./services/gameVisibilityFilters.js";
 
-/**
- * Log action types whose payloads carry private trade terms (capital/tiles and
- * the two party ids). Only the proposer and recipient may see these entries;
- * everyone else (other players + spectators) must not receive them at all —
- * mirroring how `filterTradeOffersForViewer` hides offer state itself.
- */
-const PRIVATE_TRADE_LOG_ACTIONS = new Set([
-  "trade_proposed",
-  "trade_accepted",
-  "trade_rejected",
-  "trade_expired",
-  "trade_countered",
-]);
-
-function isTradeParticipant(
-  payload: GameLogEntry["payload"],
-  viewerId: string,
-): boolean {
-  if (!payload || typeof payload !== "object") return false;
-  const { proposerId, recipientId } = payload as {
-    proposerId?: unknown;
-    recipientId?: unknown;
-  };
-  return proposerId === viewerId || recipientId === viewerId;
-}
-
-/**
- * Drop private trade log entries for viewers that are not a party to the trade.
- * `viewerId` is the player id for player views, or `null`/spectator for
- * spectators (who never participate, so all private trade entries are removed).
- */
-export function redactLogEntriesForViewer<
-  TEntry extends Pick<GameLogEntry, "actionType" | "payload">,
->(entries: TEntry[], viewerId: string | null): TEntry[] {
-  return entries.filter((entry) => {
-    if (!PRIVATE_TRADE_LOG_ACTIONS.has(entry.actionType)) return true;
-    return viewerId !== null && isTradeParticipant(entry.payload, viewerId);
-  });
-}
-
-/** Persisted `state_json` may include server-only affinity assignments. */
-export type PersistedGameState = GameState & {
-  affinityAssignments?: Record<string, string>;
-  negotiationThreads?: GameNegotiationThread[];
-  handshakeAgreements?: InGameHandshakeAgreement[];
-  tradeOffers?: TradeOffer[];
-  pendingInsiderPeek?: PendingInsiderPeek | null;
-};
-
-type ClientPendingAuction = PendingAuction & {
-  submissionCount: number;
-  mySubmission?: number | "pass";
-};
+export type {
+  BroadcastViewer,
+  PreparedGameEvent,
+  ScopableBroadcastPayload,
+  ScopableGameEvent,
+  StrippedBroadcastState,
+} from "./services/gameBroadcastVisibility.js";
+export {
+  broadcastEventStateFields,
+  buildGameScheduleEvent,
+  prepareScopableGameEvent,
+  scopeGameEventForViewer,
+  splitBroadcastPayload,
+} from "./services/gameBroadcastVisibility.js";
+// Re-exports: visibility filters and the broadcast privacy contract live in
+// focused modules now; `gameStateView.ts` is a thin orchestrator that composes
+// them into the per-viewer client view. Re-exported here so existing importers
+// (routes, persistence, tests) keep a single entry point.
+export type {
+  PersistedGameState,
+  ViewerMode,
+} from "./services/gameVisibilityFilters.js";
+export {
+  redactLogEntriesForViewer,
+  redactPendingAuctionForBroadcast,
+} from "./services/gameVisibilityFilters.js";
 
 /** HTTP/WS game state after affinity redaction and visibility filtering. */
 export type ClientGameState = Omit<
@@ -77,48 +50,6 @@ export type ClientGameState = Omit<
   pendingInsiderPeek?: PersistedGameState["pendingInsiderPeek"];
 };
 
-function withSubmissionCount(auction: PendingAuction): ClientPendingAuction {
-  return {
-    ...auction,
-    submissionCount: Object.keys(auction.submissions).length,
-  };
-}
-
-function redactPendingAuctionByMode(
-  auction: PendingAuction,
-  mode: "broadcast" | "spectator" | "player",
-  viewerId?: string,
-): ClientPendingAuction {
-  if (
-    auction.auctionType === "open_bids" ||
-    auction.auctionType === "live_bidding"
-  ) {
-    return withSubmissionCount(auction);
-  }
-
-  const submissionCount = Object.keys(auction.submissions).length;
-  if (mode !== "player") {
-    const { submissions: _submissions, ...rest } = auction;
-    return { ...rest, submissions: {}, submissionCount };
-  }
-
-  if (!viewerId) {
-    throw new Error("viewerId is required for player auction redaction");
-  }
-  const mySubmission = auction.submissions[viewerId];
-  return {
-    ...auction,
-    submissions: {},
-    submissionCount,
-    ...(mySubmission !== undefined ? { mySubmission } : {}),
-  };
-}
-
-export function redactPendingAuctionForBroadcast(
-  auction: PendingAuction,
-): ClientPendingAuction {
-  return redactPendingAuctionByMode(auction, "broadcast");
-}
 function buildClientGameStateBase(
   state: PersistedGameState,
   extras: {
@@ -166,7 +97,7 @@ function buildClientGameStateBase(
  */
 export function toClientGameStateFromInternal(
   state: InternalGameState,
-  mode: "spectator" | "player",
+  mode: ViewerMode,
   playerId: string,
 ): ClientGameState {
   return toClientGameState(
@@ -177,12 +108,16 @@ export function toClientGameStateFromInternal(
 }
 
 /**
- * Strip hidden affinity data for HTTP responses.
+ * Compile the per-viewer client view: redact each privacy-sensitive field
+ * (auction submissions, negotiation threads, handshakes, trade offers, insider
+ * peek, affinity) down to the viewer's slice via the per-domain filters in
+ * `gameVisibilityFilters.ts`, then assemble the wire shape.
+ *
  * Callers must enforce authZ (player vs spectator) before using this.
  */
 export function toClientGameState(
   state: PersistedGameState,
-  mode: "spectator" | "player",
+  mode: ViewerMode,
   playerId: string,
 ): ClientGameState {
   const pendingAuction = state.pendingAuction
@@ -226,169 +161,4 @@ export function toClientGameState(
       ? { myAffinityCardId, pendingInsiderPeek: insiderPeek }
       : {}),
   });
-}
-
-/**
- * Shared party/spectator visibility filter. Returns the input untouched when it
- * is empty/undefined; otherwise keeps each item that `isVisible(item, viewerId,
- * mode)` accepts. Each item's predicate decides whether spectators see it (so
- * e.g. open negotiation threads stay visible to spectators while private trade
- * offers/handshakes are hidden).
- */
-function filterVisibleToViewer<TItem>(
-  items: TItem[] | undefined,
-  viewerId: string,
-  mode: "spectator" | "player",
-  isVisible: (
-    item: TItem,
-    viewerId: string,
-    mode: "spectator" | "player",
-  ) => boolean,
-): TItem[] | undefined {
-  if (!items?.length) return items;
-  return items.filter((item) => isVisible(item, viewerId, mode));
-}
-
-function filterTradeOffersForViewer(
-  offers: PersistedGameState["tradeOffers"],
-  viewerId: string,
-  mode: "spectator" | "player",
-) {
-  return filterVisibleToViewer(offers, viewerId, mode, (offer, id, m) =>
-    m === "spectator"
-      ? false
-      : offer.proposerId === id || offer.recipientId === id,
-  );
-}
-
-function filterNegotiationThreadsForViewer(
-  threads: PersistedGameState["negotiationThreads"],
-  viewerId: string,
-  mode: "spectator" | "player",
-) {
-  return filterVisibleToViewer(threads, viewerId, mode, (thread, id, m) =>
-    m === "spectator"
-      ? thread.visibility === "open"
-      : thread.visibility === "open" || thread.partyIds.includes(id),
-  );
-}
-
-function filterHandshakesForViewer(
-  handshakes: PersistedGameState["handshakeAgreements"],
-  viewerId: string,
-  mode: "spectator" | "player",
-) {
-  return filterVisibleToViewer(handshakes, viewerId, mode, (entry, id, m) =>
-    m === "spectator" ? false : entry.partyA === id || entry.partyB === id,
-  );
-}
-
-/** Identifies the WebSocket session a broadcast event is being scoped for. */
-export interface BroadcastViewer {
-  viewerId: string;
-  spectator: boolean;
-}
-
-/**
- * A realtime game event as it travels on the wire BEFORE per-viewer scoping. The
- * broadcast source strips private `tradeOffers` terms off `state` and carries
- * them on the separate `tradeOffers` field (see `prepareGameBroadcastPayload`);
- * `scopeGameEventForViewer` re-injects them per viewer and redacts everything
- * down to the viewer's own slice. Extra event fields (type, sentAt, gameId, …)
- * are preserved untouched.
- */
-export interface ScopableGameEvent {
-  state: Record<string, unknown>;
-  /**
-   * The side-channel array of EVERY party's offers (re-injected into `state` by
-   * `prepareScopableGameEvent` before normalization). Typed as the offer array
-   * rather than `unknown` since that is the only payload the broadcast carries
-   * here; `prepareScopableGameEvent` still guards with `Array.isArray`.
-   */
-  tradeOffers?: TradeOffer[];
-  logEntries?: Array<Pick<GameLogEntry, "actionType" | "payload">>;
-  // Broadcast events carry extra transport fields (type, gameId, sentAt, …) that
-  // pass through untouched, so an open index signature is intentional here.
-  [key: string]: unknown;
-}
-
-/**
- * A broadcast event whose `state` has already been normalized exactly ONCE (with
- * the carried trade offers re-injected). Produced by `prepareScopableGameEvent`
- * and consumed per-viewer by `scopeGameEventForViewer`, so the per-viewer step is
- * a pure redaction that never mutates shared state.
- */
-export interface PreparedGameEvent {
-  /** Normalized engine state shared (read-only) across all viewer redactions. */
-  normalizedState: InternalGameState;
-  /** Original event fields minus `state`/`tradeOffers` (already consumed). */
-  rest: Record<string, unknown>;
-  logEntries?: Array<Pick<GameLogEntry, "actionType" | "payload">>;
-}
-
-/**
- * Per-broadcast preparation (run ONCE, before the per-viewer fan-out). Re-injects
- * the separately-carried `tradeOffers` into `state` and normalizes it a single
- * time. `normalizeGameState` mutates its argument, so doing this once here — and
- * sharing the resulting object read-only with every per-viewer redaction —
- * guarantees no shared nested mutation can leak across viewers (each viewer's
- * `scopeGameEventForViewer` only reads from it).
- *
- * The side-channel `tradeOffers` (which holds EVERY party's terms) is dropped
- * from the carried `rest` here — each viewer's own offers are re-derived, redacted,
- * inside their `scopedState.tradeOffers`. Keeping the raw array would leak foreign
- * terms to every recipient.
- */
-export function prepareScopableGameEvent(
-  event: ScopableGameEvent,
-): PreparedGameEvent {
-  // Re-inject the separately-carried trade offers so the per-viewer filter can
-  // keep each viewer's own slice. Falls back to any `tradeOffers` already on
-  // `state` for legacy/other callers.
-  const carriedTradeOffers = Array.isArray(event.tradeOffers)
-    ? event.tradeOffers
-    : null;
-  const rawState = carriedTradeOffers
-    ? { ...event.state, tradeOffers: carriedTradeOffers }
-    : event.state;
-  const { tradeOffers: _carried, state: _state, ...rest } = event;
-  return {
-    normalizedState: normalizeGameState(rawState),
-    rest,
-    logEntries: event.logEntries,
-  };
-}
-
-/**
- * Single per-viewer scoping algorithm for game broadcasts. Operates on the
- * already-normalized state from `prepareScopableGameEvent` (so this step performs
- * NO normalization or mutation — it is a pure redaction). Runs
- * `toClientGameStateFromInternal` (which redacts trade offers / auctions /
- * handshakes / negotiation threads / affinity to the viewer's slice) and redacts
- * the log entries for the viewer. Returns a new event with `state` (and
- * `logEntries`, when present) replaced by the scoped versions; all other event
- * fields pass through unchanged. This is the ONE implementation of the DO
- * fan-out scoping so the transport layer never re-derives the redaction.
- */
-export function scopeGameEventForViewer(
-  prepared: PreparedGameEvent,
-  viewer: BroadcastViewer,
-): Record<string, unknown> {
-  const scopedState = toClientGameStateFromInternal(
-    prepared.normalizedState,
-    viewer.spectator ? "spectator" : "player",
-    viewer.viewerId,
-  );
-  const scopedLogEntries = prepared.logEntries
-    ? redactLogEntriesForViewer(
-        prepared.logEntries,
-        viewer.spectator ? null : viewer.viewerId,
-      )
-    : undefined;
-
-  return {
-    ...prepared.rest,
-    state: scopedState,
-    ...(scopedLogEntries ? { logEntries: scopedLogEntries } : {}),
-  };
 }
